+++ Makie/test/runtests.jl
@@ -53,6 +53,10 @@ end
         include("SceneLike/PolarAxis.jl")
     end
 
+    @testset "GUI" begin
+        include("gui.jl")
+    end
+
     @testset "Conversion & Projection Pipeline" begin
         # TODO: move some things in here
         include("conversions/convert_arguments.jl")